+++ src/xurrent/people.py
@@ -147,4 +147,55 @@ class Person(JsonSerializableDict):
         uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/restore'
         return self._connection_object.api_call(uri, 'POST')
 
+    def get_cis(self) -> List:
+        """Retrieve configuration items for this person instance."""
+        from .configuration_items import ConfigurationItem
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/cis'
+        response = self._connection_object.api_call(uri, 'GET')
+        return [ConfigurationItem.from_data(self._connection_object, ci) for ci in response]
+
+    def get_addresses(self) -> List[dict]:
+        """Retrieve addresses for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/addresses'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_contacts(self) -> List[dict]:
+        """Retrieve contact information for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/contacts'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_permissions(self) -> List[dict]:
+        """Retrieve permissions for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/permissions'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_ci_coverages(self) -> List[dict]:
+        """Retrieve CI coverages for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/ci_coverages'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_sla_coverages(self) -> List[dict]:
+        """Retrieve SLA coverages for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/sla_coverages'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_service_coverages(self) -> List[dict]:
+        """Retrieve service coverages for this person instance."""
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/service_coverages'
+        return self._connection_object.api_call(uri, 'GET')
+
+    def get_out_of_office_periods(self) -> List:
+        """Retrieve out-of-office periods for this person instance."""
+        from .out_of_office_periods import OutOfOfficePeriod
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/out_of_office_periods'
+        response = self._connection_object.api_call(uri, 'GET')
+        return [OutOfOfficePeriod.from_data(self._connection_object, p) for p in response]
+
+    def get_skill_pools(self) -> List:
+        """Retrieve skill pools for this person instance."""
+        from .skill_pools import SkillPool
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/skill_pools'
+        response = self._connection_object.api_call(uri, 'GET')
+        return [SkillPool.from_data(self._connection_object, sp) for sp in response]
+