+++ src/xurrent/teams.py
@@ -113,3 +113,10 @@ class Team(JsonSerializableDict):
         """
         uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/trash'
         return self._connection_object.api_call(uri, 'POST')
+
+    def get_service_instances(self) -> List:
+        """Retrieve service instances assigned to this team."""
+        from .service_instances import ServiceInstance
+        uri = f'{self._connection_object.base_url}/{self.__resourceUrl__}/{self.id}/service_instances'
+        response = self._connection_object.api_call(uri, 'GET')
+        return [ServiceInstance.from_data(self._connection_object, si) for si in response]